% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{age_in_range}
\alias{age_in_range}
\title{Computes age from startdate and returns a logical vector indicating whether
the age is between lwr and upr years}
\usage{
age_in_range(startdate, lwr, upr)
}
\arguments{
\item{startdate}{Date-valued vector}

\item{lwr}{lower bound on age in years}

\item{upr}{upper bound on age in years}
}
\description{
Computes age from startdate and returns a logical vector indicating whether
the age is between lwr and upr years
}
