% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-assert.R
\name{assert_logical}
\alias{assert_logical}
\title{Assertion for logical type}
\usage{
assert_logical(vec, expr_chr)
}
\arguments{
\item{vec}{vector to check}

\item{expr_chr}{Character representation of generating expression, for messaging}
}
\description{
Assertion for logical type
}
