% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-assert.R
\name{assert_no_missing}
\alias{assert_no_missing}
\title{Assertion for non-missing values}
\usage{
assert_no_missing(vec)
}
\arguments{
\item{vec}{vector to check}
}
\description{
Assertion for non-missing values
}
