% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aux_info}
\alias{aux_info}
\title{Auxiliary information for performing rule checks}
\format{
a list with (at last count) 24 elements:
\describe{
\item{valid_s_reg_statuses}{From the google doc shared by Justin}
\item{valid_c_line_items}{From the google doc shared by Justin}
\item{valid_instruction_method_codes}{From the google doc shared by Justin}
\item{valid_program_types}{From the google doc shared by Justin}
\item{valid_instruct_types}{From the google doc shared by Justin}
\item{valid_cip_codes}{From the CIPCode2020.csv in sandbox/, from Justin}
\item{valid_highschools}{From sandbox/highschools.txt (pipe-separated) - HS_ACT_CODE}
\item{non_concurrent_highschools}{Pulled from sql code in data inventory for SC12b}
\item{concurrent_course_ids}{from sandbox/analytics_quad_concurrent_courses.csv,
but should change in future (TODO)}
\item{ut_highschools}{From sandbox/highschools.txt}
\item{valid_degree_ids}{From Reference.dbo.Degree_type.xlsx, now in "sandbox/degree-types.csv"}
\item{valid_previous_degree_types}{From the google doc shared by Justin}
\item{valid_level_class_ids}{From the google doc shared by Justin (s_levels)}
\item{valid_seasons}{Pulled from sql in G25a row of data inventory xlsx}
\item{valid_final_grades}{From the google doc shared by Justin}
\item{passing_grades}{From the sql in SC08d in data inventory}
\item{valid_building_location_codes}{From the google doc shared by Justin}
\item{valid_ownership_codes}{From the google doc shared by Justin}
\item{valid_building_condition_codes}{From the google doc shared by Justin}
\item{valid_room_group1_codes}{From the google doc shared by Justin}
\item{valid_room_use_code_groups}{From the google doc shared by Justin}
\item{valid_room_use_codes}{From the google doc shared by Justin}
\item{building_inventory}{From sandbox/analytics_quad_buildings.csv, but should change (TODO)}
\item{rooms_inventory}{From sandbox/analytics_quad_buildings.csv, but should change (TODO)}

}
}
\source{
Defined in data-raw/aux_info.R
}
\usage{
aux_info
}
\description{
While the \code{checklist} object contains the expressions to perform each rule check,
there may be additional information required beyond what is defined in the
\code{checker} expression and the dataframe being checked. The \code{aux_info} provides
this information as a list, primarily of vectors of valid values.
}
\keyword{datasets}
