% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{check_expected_values}
\alias{check_expected_values}
\title{QC-check on expected values supplied in testdata dataframe (from csv)}
\usage{
check_expected_values(
  df,
  colname = "Expected value",
  valid_values = c("pass", "fail")
)
}
\arguments{
\item{df}{dataframe to check}

\item{colname}{Column name containing expected values}

\item{valid_values}{which values should be considered valid?}
}
\description{
Removes rows with a warning when column values are not in \code{valid_values}
}
