% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth-tables.R
\name{combine_same_vars}
\alias{combine_same_vars}
\title{Utility function to combine same-name elements of veclist to produce a list
with unique names and unique values per element}
\usage{
combine_same_vars(veclist)
}
\arguments{
\item{veclist}{a list of vectors}
}
\description{
Utility function to combine same-name elements of veclist to produce a list
with unique names and unique values per element
}
