% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-output.R
\name{compare_rule_output}
\alias{compare_rule_output}
\title{Show comparison between expected and obtained rule output, along with helpful info}
\usage{
compare_rule_output(
  rule,
  testdf = get_test_data("student"),
  mismatch_only = TRUE
)
}
\arguments{
\item{rule}{character name of rule, e.g. "S17b"}

\item{testdf}{dataframe with test data, like that found in extdata/student_unit_test.csv}

\item{mismatch_only}{if TRUE (default), only show rows where expected rule outcome does not match actual}
}
\description{
Show comparison between expected and obtained rule output, along with helpful info
}
