% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{course_conditional_check}
\alias{course_conditional_check}
\title{Function for checking conditional missingness for rules like C18a, C19a}
\usage{
course_conditional_check(
  x,
  instruction_method_code,
  section_format_type_code,
  budget_code,
  campus_id
)
}
\arguments{
\item{x}{vector of values to check, e.g. from building_name or building_number}

\item{instruction_method_code, section_format_type_code, budget_code, campus_id}{relevant vector columns of df_tocheck}
}
\description{
Function for checking conditional missingness for rules like C18a, C19a
}
