% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{equivalent}
\alias{equivalent}
\title{NA-imputing double-equals replacement}
\usage{
equivalent(x, y)
}
\arguments{
\item{x, y}{logical vectors}
}
\description{
Returns TRUE if x and y are equal OR both are missing, otherwise FALSE
}
