% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_buildings_validation}
\alias{fake_buildings_validation}
\title{Fake buildings validation file}
\format{
A data frame with 112 rows and 13 variables:
\describe{
\item{building_activity_date}{}
\item{building_area_gross}{}
\item{building_auxiliary}{}
\item{building_auxiliary_12}{}
\item{building_condition_code}{}
\item{building_condition_desc}{}
\item{building_construction_year}{}
\item{building_cost_myr}{}
\item{building_cost_replacement}{}
\item{building_location_code}{}
\item{building_location_desc}{}
\item{building_number}{}
\item{building_remodel_year}{}
}
}
\source{
The script for creating this data set is in the dev folder
}
\usage{
fake_buildings_validation
}
\description{
A dataset containing simulated data for the buildings validation process
with USHE. This dataset would be pulled from the data warehouse, and then fed into
the validation scripts in the utValidateR package.
}
\keyword{datasets}
