% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_course_validation}
\alias{fake_course_validation}
\title{Fake course validation file}
\format{
A data frame with 5,132 rows and 49 variables:
\describe{
\item{academic_department_id}{}
\item{budget_code}{}
\item{building_number_1}{}
\item{building_number_2}{}
\item{building_number_3}{}
\item{campus_id}{}
\item{class_size}{}
\item{college_id}{}
\item{contact_hours}{}
\item{course_max_credits}{}
\item{course_min_credits}{}
\item{course_number}{}
\item{course_reference_number}{}
\item{course_title}{}
\item{instruction_method_code}{}
\item{instructor_employee_id}{}
\item{instructor_name}{}
\item{meet_building_id_1}{}
\item{meet_building_id_2}{}
\item{meet_building_id_3}{}
\item{meet_days_1}{}
\item{meet_days_2}{}
\item{meet_days_3}{}
\item{meet_end_date}{}
\item{meet_end_time_1}{}
\item{meet_end_time_2}{}
\item{meet_end_time_3}{}
\item{meet_room_number_1}{}
\item{meet_room_number_2}{}
\item{meet_room_number_3}{}
\item{meet_start_time_1}{}
\item{meet_start_time_2}{}
\item{meet_start_time_3}{}
\item{program_type}{}
\item{room_max_occupancy_1}{}
\item{room_max_occupancy_2}{}
\item{room_max_occupancy_3}{}
\item{room_use_code_1}{}
\item{room_use_code_2}{}
\item{room_use_code_3}{}
\item{scbcrse_activity_date}{}
\item{section_format_type_code}{}
\item{section_number}{}
\item{sirasgn_activity_date}{}
\item{spriden_activity_date}{}
\item{ssbsect_activity_date}{}
\item{ssrmeet_activity_date}{}
\item{subject_code}{}
\item{term_id}{}
}
}
\source{
The script for creating this data set is in the dev folder
}
\usage{
fake_course_validation
}
\description{
A dataset containing simulated data for the course validation process
with USHE. This dataset would be pulled from the data warehouse, and then fed into
the validation scripts in the utValidateR package.
}
\keyword{datasets}
