% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_graduation_validation}
\alias{fake_graduation_validation}
\title{Fake graduation validation file}
\format{
A data frame with 13,866 rows and 44 variables:
\describe{
\item{birth_date}{}
\item{cumulative_graduation_gpa}{}
\item{degree_desc}{}
\item{degree_id}{}
\item{ethnicity_code}{}
\item{ethnicity_desc}{}
\item{first_admit_county_code}{}
\item{first_name}{}
\item{gender_code}{}
\item{gorsdav_activity_date}{}
\item{graduated_academic_year_code}{}
\item{graduated_term_id}{}
\item{graduation_date}{}
\item{high_school_code}{}
\item{ipeds_award_level_code}{}
\item{ipeds_race_ethnicity}{}
\item{is_american_indian_alaskan}{}
\item{is_asian}{}
\item{is_black}{}
\item{is_hawaiian_pacific_islander}{}
\item{is_hispanic_latino_ethnicity}{}
\item{is_international}{}
\item{is_other_race}{}
\item{is_white}{}
\item{last_name}{}
\item{middle_name}{}
\item{name_suffix}{}
\item{overall_cumulative_credits_earned}{}
\item{previous_degree_type}{}
\item{primary_major_cip_code}{}
\item{primary_major_college_desc}{}
\item{primary_major_desc}{}
\item{primary_program_id}{}
\item{required_credits}{}
\item{season}{}
\item{shrdgmr_activity_date}{}
\item{shrtgpa_activity_date}{}
\item{sis_student_id}{}
\item{sis_system_id}{}
\item{sorhsch_activity_date}{}
\item{ssn}{}
\item{total_cumulative_ap_credits_earned}{}
\item{total_cumulative_clep_credits_earned}{}
\item{total_cumulative_credits_attempted_other_sources}{}
\item{total_remedial_hours}{}
\item{transfer_cumulative_credits_earned}{}
}
}
\source{
The script for creating this data set is in the dev folder
}
\usage{
fake_graduation_validation
}
\description{
A dataset containing simulated data for the graduation validation process
with USHE. This dataset would be pulled from the data warehouse, and then fed into
the validation scripts in the utValidateR package.
}
\keyword{datasets}
