% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_rooms_validation}
\alias{fake_rooms_validation}
\title{Fake rooms validation file}
\format{
A data frame with 7,102 rows and 10 variables:
\describe{
\item{building_activity_date}{}
\item{building_construction_year}{}
\item{building_number}{}
\item{buildings_id}{}
\item{room_activity_date}{}
\item{room_area}{}
\item{room_disabled_access}{}
\item{room_group1_code}{}
\item{room_name}{}
\item{room_number}{}
\item{room_prorated}{}
\item{room_prorated_area}{}
\item{room_stations}{}
\item{room_use_code}{}
\item{room_use_code_group}{}
\item{rooms_id}{}
}
}
\source{
The script for creating this data set is in the dev folder
}
\usage{
fake_rooms_validation
}
\description{
A dataset containing simulated data for the rooms validation process
with USHE. This dataset would be pulled from the data warehouse, and then fed into
the validation scripts in the utValidateR package.
}
\keyword{datasets}
