% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_student_course_validation}
\alias{fake_student_course_validation}
\title{Fake student-courses validation file}
\format{
A data frame with 91,570 rows and 18 variables:
\describe{
\item{attempted_credits}{}
\item{attribute_code}{}
\item{budget_code}{}
\item{contact_hours}{}
\item{course_level_id}{}
\item{course_number}{}
\item{course_reference_number}{}
\item{earned_credits}{}
\item{final_grade}{}
\item{latest_high_school_code}{}
\item{part_term_weeks}{}
\item{section_number}{}
\item{sis_student_id}{}
\item{sis_system_id}{}
\item{ssbsect_activity_date}{}
\item{ssn}{}
\item{subject_code}{}
\item{term_id}{}
}
}
\source{
The script for creating this data set is in the dev folder
}
\usage{
fake_student_course_validation
}
\description{
A dataset containing simulated data for the student-courses validation process
with USHE. This dataset would be pulled from the data warehouse, and then fed into
the validation scripts in the utValidateR package.
}
\keyword{datasets}
