% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_checks.R
\name{get_activity_dates}
\alias{get_activity_dates}
\title{Returns a dataframe with columns (rule)_activity_date, (rule)_error_age}
\usage{
get_activity_dates(df, datecol, rule)
}
\arguments{
\item{df}{the dataframe to check}

\item{datecol}{Name of the relevant activity_date column}

\item{rule}{Name of the rule, e.g "S03b"}
}
\description{
Returns a dataframe with columns (rule)_activity_date, (rule)_error_age
}
