% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{get_col_spec}
\alias{get_col_spec}
\title{Returns the spec (output of \code{readr::cols()}) for the specified file}
\usage{
get_col_spec(file = c("student"))
}
\arguments{
\item{file}{e.g. "student" (actually that's the only currently valid value)}
}
\description{
Returns the spec (output of \code{readr::cols()}) for the specified file
}
