% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{get_test_data}
\alias{get_test_data}
\title{Returns a dataframe containing info needed for \code{do_checks()} plus expected outcome for specified rules}
\usage{
get_test_data(file = c("student"), checklist)
}
\arguments{
\item{file}{which file to get. Currently only "student" is available}

\item{checklist}{used for checking rule names (see \code{check_rule_names()})}
}
\description{
This data is specified by UT Tech, and located in inst/testdata/
}
