% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{get_test_data_csv_loc}
\alias{get_test_data_csv_loc}
\title{Returns the location of the csv containing test data (with expected outcomes) for the given file}
\usage{
get_test_data_csv_loc(file = c("student"))
}
\arguments{
\item{file}{e.g. "student" (actually that's the only currently valid value)}
}
\description{
Returns the location of the csv containing test data (with expected outcomes) for the given file
}
