% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth-tables.R
\name{get_truth_table}
\alias{get_truth_table}
\title{High-level truth-table generator, returns a dataframe with all variables in
an expression and a logical \code{passes} column that is the result of calling the
expression on the dataframe}
\usage{
get_truth_table(x, limit = Inf, extra_rows = NULL)
}
\arguments{
\item{x}{an expression like those in \code{checklist$checker}}

\item{limit}{maximum number of example rows to return per truth outcome (T or F)}

\item{extra_rows}{optional dataframe to which the expression will be applied,
supplementing rows of returned truth table}
}
\description{
High-level truth-table generator, returns a dataframe with all variables in
an expression and a logical \code{passes} column that is the result of calling the
expression on the dataframe
}
