% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{get_tryerror}
\alias{get_tryerror}
\title{Returns the text of the error message in a \code{try()} failure}
\usage{
get_tryerror(tryerror)
}
\arguments{
\item{tryerror}{result of \code{try()} failure}
}
\description{
Returns the text of the error message in a \code{try()} failure
}
