% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth-tables.R
\name{get_tt_vecs}
\alias{get_tt_vecs}
\title{Should return a named list of vectors, which can be supplied to \code{expand.grid()}}
\usage{
get_tt_vecs(x, aux_info)
}
\arguments{
\item{x}{an expression like those in \code{checklist$checker}}

\item{aux_info}{a list of auxiliary information, as used in \code{do_checks()}}
}
\description{
Recursive along sub-expressions of x, calls \code{get_tt_vecs_basecase} in base-case
}
