% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth-tables.R
\name{get_tt_vecs_basecase}
\alias{get_tt_vecs_basecase}
\title{Returns a list of example values for creating truth table}
\usage{
get_tt_vecs_basecase(base_expr, aux_info)
}
\arguments{
\item{base_expr}{A non-nested expression}
}
\description{
Might make more sense to code this directly into if() clause in get_tt_vecs()
}
