% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{has_nonalpha}
\alias{has_nonalpha}
\title{Returns TRUE for elements of x with non-alphabet characters; FALSE otherwise}
\usage{
has_nonalpha(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
Returns TRUE for elements of x with non-alphabet characters; FALSE otherwise
}
