% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{in_range}
\alias{in_range}
\title{Checks whether a number is between lwr and upr}
\usage{
in_range(x, lwr, upr)
}
\arguments{
\item{x}{Numeric vector}

\item{lwr, upr}{Limits on x, inclusive}
}
\description{
Returns TRUE for x between lwr and upr, FALSE otherwise. Missing maps to FALSE.
}
