% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_alpha_chr}
\alias{is_alpha_chr}
\title{Checks that a character vector is alpha-only (plus apostrophe)}
\usage{
is_alpha_chr(x, missing_ok = TRUE)
}
\arguments{
\item{x}{Character vector}

\item{missing_ok}{if TRUE (default), do not flag \code{NA} or \code{""} values}
}
\description{
Checks that a character vector is alpha-only (plus apostrophe)
}
