% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth-tables.R
\name{is_base_expression}
\alias{is_base_expression}
\title{Function for checking whether to treat as base-case or continue recursion}
\usage{
is_base_expression(x)
}
\description{
Function for checking whether to treat as base-case or continue recursion
}
