% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_duplicated}
\alias{is_duplicated}
\title{Returns TRUE for duplicated elements of x; FALSE otherwise}
\usage{
is_duplicated(x, count_missing = FALSE)
}
\arguments{
\item{x}{a vector}

\item{count_missing}{if TRUE, count multiple missing values as duplicates,
otherwise treat as incomparable. Defaults to FALSE.}
}
\description{
Returns TRUE for duplicated elements of x; FALSE otherwise
}
