% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_missing_chr}
\alias{is_missing_chr}
\title{Checker for missing character values}
\usage{
is_missing_chr(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
Logical vector
}
\description{
Checker for missing character values
}
\section{Functions}{
\itemize{
\item \code{is_missing_chr()}: returns TRUE for NA or empty-string values

}}
