% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_undergrad_level}
\alias{is_undergrad_level}
\title{Returns TRUE if level is one of "FR", "SO", "JR", "SR", and FALSE otherwise}
\usage{
is_undergrad_level(level)
}
\arguments{
\item{level}{character vector}
}
\description{
Returns TRUE if level is one of "FR", "SO", "JR", "SR", and FALSE otherwise
}
