% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_utah_county}
\alias{is_utah_county}
\alias{is_us_county}
\alias{is_us_state}
\title{Checks whether a county is in Utah.}
\usage{
is_utah_county(county_code)

is_us_county(county_code)

is_us_state(state)
}
\arguments{
\item{county_code}{Vector of first_admit_county_code values}

\item{state}{first_admit_state_code}
}
\description{
Checks whether a county is in Utah.
}
\section{Functions}{
\itemize{
\item \code{is_us_county}: Checks whether a county is in USA

\item \code{is_us_state}: Checks whether a state code is in USA
}}

