% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_valid_act_score}
\alias{is_valid_act_score}
\alias{is_valid_ssn}
\alias{is_valid_zip_code}
\alias{is_valid_ssid}
\alias{is_valid_year}
\alias{is_valid_student_id}
\alias{is_valid_previous_id}
\alias{is_valid_country_code}
\alias{is_valid_credits}
\alias{is_valid_credits_chr}
\alias{is_valid_gpa}
\alias{is_valid_student_type}
\alias{is_valid_occupancy}
\alias{is_valid_values}
\alias{is_valid_course_reference_number}
\alias{is_valid_graduation_date}
\title{Validity checks on various values}
\usage{
is_valid_act_score(x)

is_valid_ssn(x, missing_ok = TRUE)

is_valid_zip_code(x, missing_ok = TRUE)

is_valid_ssid(x)

is_valid_year(x)

is_valid_student_id(x)

is_valid_previous_id(x)

is_valid_country_code(x)

is_valid_credits(x)

is_valid_credits_chr(x)

is_valid_gpa(x)

is_valid_student_type(x)

is_valid_occupancy(x)

is_valid_values(x, valid_values, missing_ok = TRUE)

is_valid_course_reference_number(x)

is_valid_graduation_date(x)
}
\arguments{
\item{x}{Vector to check}

\item{missing_ok}{if TRUE (default), do not flag \code{NA} or \code{""} values}

\item{valid_values}{vector of valid values x can take}
}
\description{
Validity checks on various values
}
\section{Functions}{
\itemize{
\item \code{is_valid_ssn}: ssn

\item \code{is_valid_zip_code}: zip_code

\item \code{is_valid_ssid}: SSID

\item \code{is_valid_year}: 4-digit year in 2000-2099 as character string

\item \code{is_valid_student_id}: student_id

\item \code{is_valid_previous_id}: previous_student_id

\item \code{is_valid_country_code}: first_admit_country_code

\item \code{is_valid_credits}: credits_earned

\item \code{is_valid_credits_chr}: character-valued credits (min_credits and max_credits)

\item \code{is_valid_gpa}: gpa

\item \code{is_valid_student_type}: student_type_code

\item \code{is_valid_occupancy}: room occupancy

\item \code{is_valid_values}: matching a vector of values

\item \code{is_valid_course_reference_number}: course reference number

\item \code{is_valid_graduation_date}: graduation date
}}

