% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_valid_dates_for_term}
\alias{is_valid_dates_for_term}
\title{Helper function for valid date for term}
\usage{
is_valid_dates_for_term(date_, term_id, term_sufx, campus_id)
}
\arguments{
\item{date_}{Date would be either start term or end term}

\item{term_id}{It is the year with suffix}

\item{term_sufx}{It is the suffix only}

\item{campus_id}{valid campus id}
}
\description{
Helper function for valid date for term
}
