% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_valid_pf_degree_level_match}
\alias{is_valid_pf_degree_level_match}
\title{validate degree level mapping for programs audit}
\usage{
is_valid_pf_degree_level_match(ipeds_award_level_code, degree_level)
}
\arguments{
\item{ipeds_award_level_code}{IPEDS award level code}

\item{degree_level}{Degree level associated with the degree type}
}
\value{
Logical vector
}
\description{
\code{is_valid_pf_degree_level_match} Validates degree type aligns with degree level
}
