% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth-tables.R
\name{lookup_tt_vecs}
\alias{lookup_tt_vecs}
\title{Returns a named list containing vector of test values}
\usage{
lookup_tt_vecs(fun, args, aux_info)
}
\arguments{
\item{fun}{character function name}

\item{args}{Named list of arguments}

\item{aux_info}{a list with auxiliary information to be used by the expression}
}
\description{
Returns a named list containing vector of test values
}
