% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{make_checker}
\alias{make_checker}
\title{Function factory for generating checker functions}
\usage{
make_checker(rule, expr)
}
\arguments{
\item{rule}{Name of the rule being applied, e.g. "S03b"}

\item{expr}{Expression that generates a logical vector when applied to a dataframe}
}
\description{
Result is a function with a single dataframe argument
}
