% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{matches_regex}
\alias{matches_regex}
\title{Helper function for regex matching with missingness toggle}
\usage{
matches_regex(x, regex, missing_ok)
}
\arguments{
\item{x}{Vector to check}

\item{regex}{regular expression string}

\item{missing_ok}{if TRUE (default), do not flag \code{NA} or \code{""} values}
}
\description{
Helper function for regex matching with missingness toggle
}
