% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_missing_space}
\alias{is_missing_space}
\title{Helper functions for missing building numbers}
\usage{
is_missing_space(space, start_time, meet_days)
}
\arguments{
\item{space}{vector of space i.e building_id,}

\item{start_time}{vector of start_time values}

\item{meet_days}{vector of meet_day values}
}
\description{
Helper functions for missing building numbers
}
\section{Functions}{
\itemize{
\item \code{is_missing_space()}: Building check whether a time or day exist and if the building number is missing

}}
