% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\docType{data}
\name{ssn_regex}
\alias{ssn_regex}
\title{Regex to check ssn, from:
https://www.geeksforgeeks.org/how-to-validate-ssn-social-security-number-using-regular-expression}
\format{
An object of class \code{character} of length 1.
}
\usage{
ssn_regex
}
\description{
Regex to check ssn, from:
https://www.geeksforgeeks.org/how-to-validate-ssn-social-security-number-using-regular-expression
}
\keyword{datasets}
