% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{TODO}
\alias{TODO}
\title{Helper for troubleshooting and communicating known needs}
\usage{
TODO(msg)
}
\arguments{
\item{msg}{Message to be displayed}
}
\description{
Primarily intended to aid in the development of this package
}
