% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_checks.R
\name{vars_available}
\alias{vars_available}
\title{For a single rule, returns TRUE if all needed_vars are present in avail_vars,
FALSE otherwise.}
\usage{
vars_available(needed_vars, avail_vars, rule, verbose = TRUE)
}
\arguments{
\item{needed_vars, avail_vars}{character vectors}

\item{rule}{character specifying the rule name}

\item{verbose}{if TRUE, print which variables are missing}
}
\description{
For a single rule, returns TRUE if all needed_vars are present in avail_vars,
FALSE otherwise.
}
