% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_checks.R
\name{verify_columns_present}
\alias{verify_columns_present}
\title{Returns checklist with non-applicable rules removed.}
\usage{
verify_columns_present(df_tocheck, checklist, aux_info, verbose = TRUE)
}
\arguments{
\item{df_tocheck}{a data.frame to check, like the included \code{fake_student_df}}

\item{checklist}{a tibble containing rule functions in \code{checker} column, as
well as \code{rule} and \code{activity_date} columns}

\item{aux_info}{list (or environment) with additional objects used by the checkers}

\item{verbose}{if TRUE, print additional information (currently only passed to
\code{verify_columns_present()})}
}
\description{
A rule is non-applicable if the information it requires (via \code{all.vars()}) is
not present in either \code{df_tocheck} or \code{aux_info}.
}
