% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-assert.R
\name{assert_length}
\alias{assert_length}
\title{Assertion for correct vector length}
\usage{
assert_length(vec, df)
}
\arguments{
\item{vec}{vector to check}

\item{df}{dataframe being checked}
}
\description{
Assertion for correct vector length
}
