% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{checklist}
\alias{checklist}
\title{Rule specification using R expressions}
\format{
a tibble with ?? columns and ?? variables:
\describe{
\item{rule}{Rule name, e.g. S01a}
\item{description}{Short description of rule}
\item{status}{Whether breaking the rule is a Failure or a Warning}
\item{activity_date}{Name of the relevant activity-date variable}
\item{checker}{Expression defining the rule}
}
}
\source{
Defined in data-raw/checklist.R
}
\usage{
checklist
}
\description{
TODO: fill out documentation
}
\keyword{datasets}
