% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_checks.R
\name{do_checks}
\alias{do_checks}
\title{Apply all rules in checklist to a dataframe}
\usage{
do_checks(df_tocheck, checklist, aux_info)
}
\arguments{
\item{df_tocheck}{a data.frame to check, like fake_student_df}

\item{checklist}{a tibble containing rule functions in \code{checker} column, as
well as \code{rule} and \code{activity_date} columns}

\item{aux_info}{list (or environment) with additional objects used by the checkers}
}
\description{
Returns df_tocheck with additional columns for each rule in checklist
}
