% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{get_age}
\alias{get_age}
\title{Returns number of days since each date in x}
\usage{
get_age(x)
}
\arguments{
\item{x}{a Date vector or similar}
}
\description{
Returns number of days since each date in x
}
