% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_checklist.R
\name{get_checklist}
\alias{get_checklist}
\title{Returns the checklist for a given file and type}
\usage{
get_checklist(
  file = c("student", "student course", "graduation", "course", "buildings", "rooms"),
  type = c("database", "ushe")
)
}
\arguments{
\item{file}{e.g. "student"}

\item{type}{either "database" or "ushe"}
}
\description{
Returns the checklist for a given file and type
}
