% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth-tables.R
\name{get_truth_table}
\alias{get_truth_table}
\title{High-level truth-table generator, returns a dataframe with all variables in
an expression and a logical \code{passes} column that is the result of calling the
expression on the dataframe}
\usage{
get_truth_table(x, aux_info, limit = Inf, extra_rows = NULL)
}
\arguments{
\item{x}{an expression like those in \code{checklist$checker}}
}
\description{
High-level truth-table generator, returns a dataframe with all variables in
an expression and a logical \code{passes} column that is the result of calling the
expression on the dataframe
}
