% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_alpha_chr}
\alias{is_alpha_chr}
\alias{is_missing_chr}
\title{Checks that a character vector is alpha-only (plus apostrophe)}
\usage{
is_alpha_chr(x, missing_ok = TRUE)

is_missing_chr(x)
}
\arguments{
\item{x}{Character vector}

\item{missing_ok}{if TRUE (default), do not flag \code{NA} or \code{""} values}
}
\description{
Checks that a character vector is alpha-only (plus apostrophe)

Checker for missing character values
}
\section{Functions}{
\itemize{
\item \code{is_missing_chr}: returns TRUE for NA or empty-string values
}}

