% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_duplicated}
\alias{is_duplicated}
\title{Returns TRUE for duplicated elements of x; FALSE otherwise}
\usage{
is_duplicated(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Returns TRUE for duplicated elements of x; FALSE otherwise
}
