% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{is_valid_act_score}
\alias{is_valid_act_score}
\alias{is_valid_ssn}
\alias{is_valid_zip_code}
\alias{is_valid_class_level}
\alias{is_valid_student_id}
\alias{is_valid_previous_id}
\alias{is_valid_country_code}
\alias{is_valid_credits}
\alias{is_valid_gpa}
\alias{is_valid_gender}
\alias{is_valid_student_type}
\title{Validity checks on various values}
\usage{
is_valid_act_score(x)

is_valid_ssn(x)

is_valid_zip_code(x)

is_valid_class_level(x)

is_valid_student_id(x)

is_valid_previous_id(x)

is_valid_country_code(x)

is_valid_credits(x)

is_valid_gpa(x)

is_valid_gender(x)

is_valid_student_type(x)
}
\arguments{
\item{x}{Vector to check}
}
\description{
Validity checks on various values
}
\section{Functions}{
\itemize{
\item \code{is_valid_ssn}: ssn

\item \code{is_valid_zip_code}: zip_code

\item \code{is_valid_class_level}: level_class_id

\item \code{is_valid_student_id}: student_id

\item \code{is_valid_previous_id}: previous_student_id

\item \code{is_valid_country_code}: first_admit_country_code

\item \code{is_valid_credits}: credits_earned

\item \code{is_valid_gpa}: gpa

\item \code{is_valid_gender}: gender_code, Used in rule S13 and G06a

\item \code{is_valid_student_type}: student_type_code
}}

