% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{make_checker}
\alias{make_checker}
\title{Function factory for generating checker functions}
\usage{
make_checker(rule, expr, env = rlang::caller_env())
}
\arguments{
\item{rule}{Name of the rule being applied, e.g. "S03b"}

\item{expr}{Expression that generates a logical vector when applied to a dataframe}

\item{env}{environment for evaluating checkers, passed to \code{rlang::eval_tidy()}}
}
\description{
Used internally by \code{do_checks()} to convert r expression into function that can
be used on a dataset that needs checking. Result is a function with a single
dataframe argument.
}
