% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker-helpers.R
\name{matches_regex}
\alias{matches_regex}
\title{Wrapper around stringr::str_detect()}
\usage{
matches_regex(string, pattern)
}
\arguments{
\item{string}{passed to str_detect()}

\item{pattern}{passed to str_detect()}
}
\description{
Wrapper around stringr::str_detect()
}
